Rename the `initials` package into `avatar`
This package is responsible to handling avatars, the initials are only
one possible way to implement avatars and for what I have understood a
gravatar implementation could be implemented in the future.
