+++ pkg/avatar/service_test.go
@@ -1,4 +1,4 @@
-package initials
+package avatar
 
 import (
 	"math/rand"
